refactor(channels): relocate Discord transport primitives to channels/discord/
Mechanical move of 8 source files (and their .test.ts siblings) from
src/legacy/{gateway,discord}/ to src/channels/discord/, prepping the new
channels/ home for the Discord transport:

  legacy/gateway/reply-directive       -> channels/discord/reply-directive
  legacy/gateway/dedupe                -> channels/discord/dedupe
  legacy/gateway/mention               -> channels/discord/mention
  legacy/gateway/session-keys          -> channels/discord/session-key
  legacy/discord/thread-bindings       -> channels/discord/thread-binding
  legacy/discord/discord-a2a-sink      -> channels/discord/a2a-sink
  legacy/discord/message-metadata      -> channels/discord/message-metadata
  legacy/discord/types                 -> channels/discord/types

Renamed singulars (thread-binding, session-key, a2a-sink) drop redundant
prefixes now that the directory provides the namespace.

Updated importers to point at the new paths:
  - src/legacy/discord/{discord,discord-manager,index,config}.ts
  - src/legacy/discord/{discord,config}.test.ts
  - src/agent/tools/spawn-agent.ts (DiscordA2ASink import path only)

No logic changes; typecheck, lint, and all 691 tests pass.
