+++ src/agent/tools/spawn-agent.ts
@@ -5,7 +5,7 @@ import type { AgentRuntime } from "../runtime.js";
 import { RunValidationError } from "../types.js";
 import type { RunRequest } from "../types.js";
 import { getDiscordA2AStreamContext } from "../../legacy/discord/a2a-stream-context.js";
-import { DiscordA2ASink } from "../../legacy/discord/discord-a2a-sink.js";
+import { DiscordA2ASink } from "../../channels/discord/a2a-sink.js";
 import { getAgentEndMeta } from "../pi/messages.js";
 import { createLogger } from "../../logging/logger.js";
