+++ src/channels/discord/message-metadata.test.ts
