+++ src/channels/discord/message-metadata.ts
