+++ src/channels/discord/thread-binding.ts
