+++ src/legacy/discord/discord-manager.ts
@@ -4,11 +4,11 @@
 import type {
   SessionStore,
 } from "../../sessions/types.js";
-import type { DiscordAccountConfig } from "./types.js";
+import type { DiscordAccountConfig } from "../../channels/discord/types.js";
 import type { AgentRuntime } from "../../agent/runtime.js";
 import { getDiscordToken } from "./config.js";
 import { DiscordTransport } from "./discord.js";
-import { ThreadBindingManager } from "./thread-bindings.js";
+import { ThreadBindingManager } from "../../channels/discord/thread-binding.js";
 import { createLogger } from "../../logging/logger.js";
 
 const log = createLogger("discord-manager");
